generalise handling of szip(.dll)?.a in other builds
It was set up to work when szip is unpacked from a previous build,
in whuch case the libsz.a file existed.  But this renaming is done
in the packing step so new builds have libsz.dll.a.

The need to specify the libsz path might ultimately be due to
the pkg-config version but this approach works for now.